==> analysis/0_data_summary.Rmd <==
---
title: "EcoEncontros seminar series data"
author: "Melina Leite & Júlia Barreto"
date: "`r format(Sys.time(), '%d de %B de %Y')`"
output:
  rmdformats::readthedown:
    highlight: kate
    self_contained: true
    thumbnails: false
    lightbox: true
    gallery: false
  pdf_document:
    highlight: tango
    toc: yes
---
```{r setup, echo=FALSE, warning=FALSE, message=FALSE}
library(knitr)
library(tidyverse); library(cowplot); library(patchwork)
theme_set(theme_cowplot())
library(janitor); library(lubridate)

opts_chunk$set(fig.align = 'center', warning = FALSE, message = FALSE, error = FALSE, echo=T, cache=F)
options(formatR.arrow = TRUE, width = 90, help_type = "html")
```


# EcoEncontros Seminar

Talks from EcoEncontros Seminar series at the Graduate Program of Ecology in the University of São Paulo (PPGE-USP), Brazil

See file `metadata.txt`, in folder `data` for more description and detail of the dataset.

Reading and summarising data:
```{r}
data1 <- read.table("data/presentations_PPGE_2008-2019.csv", sep=",",
                   header=T, as.is=T)
data1$date <- dmy(data1$date)
data1$year <- year(data1$date) 
skimr::skim(data1)
```


**For ALL analyses:** Excluding special events as round tables and discussions not related to a project or study presented by someone.

```{r}
IDs <- c(90, 154, 170, 211, 212, 230, 244, 250, 289, 319)
data_all <- data1 %>% filter(!id %in% IDs)
```


Including varible to indicated the affirmative actions period:
```{r}
data_all$affirm_action <- ifelse(data_all$year<2018,"before", "after")
data_all$affirm_action <- fct_relevel(data_all$affirm_action,"before", "after")
```

There were `r table(data_all$affirm_action)[1]` seminars before and  `r table(data_all$affirm_action)[2]` after the affirmative actions.


# Population data from the Graduate Programm PPGE-USP

Data from 2008-2019.

Number of students, post-docs and professors in the PPGE-USP per gender and year.

```{r}
pop <- read.table("data/pop_PPGE_2008-2019.csv", sep=",",
                   header=T, as.is=T)
pop2 <- pop %>%  
  pivot_longer(2:7, names_to=c("category","gender"), values_to="N",
               names_sep = "_") %>%
  pivot_wider(names_from = "gender", values_from="N") %>%
  mutate(propFtotal = total_F/(total_F+total_M),
         ratioFtotal = total_F/total_M,
         propFcat = F/(F+M),
         ratioFcat = F/M) %>%
  dplyr::select(year, category, total_F, total_M, propFtotal,
                ratioFtotal,F,M, propFcat, ratioFcat)
```

Over all years
```{r}
pop2 %>% group_by(category) %>% summarise(F=sum(F), M=sum(M),
                                          prop.F = sum(F)/(sum(M)+ sum(F))) %>% 
  kable(digits=2)
```

Average over all years
```{r}
pop2 %>% group_by(category) %>% summarise(mean.propF=mean(propFcat),
                                          sd.propF = sd(propFcat)) %>% 
  kable(digits=2)
```



PPGE-USP population by gender:

```{r, echo=F}
pop %>% dplyr::select(year,total_F,total_M) %>% pivot_longer(2:3, names_to = "gender", values_to = "N") %>%
ggplot(aes(x=as.factor(year), y=N, fill=gender)) + geom_col() +
   scale_fill_manual("gender", values = c("#b2abd2", "#fdb863"),
                     label=c("F", "M")) +
  xlab("")+ ylab("Population size")+
  theme(text=element_text(size=18),
    axis.text.x = element_text(size=16,angle = 45, hjust=1)) +
  ggplot2::annotate("text", x=1:12, y= pop$total_M + pop$total_F/2, size=5,
           label = paste0(round(pop$total_F*100/(pop$total_M+pop$total_F),0),"%"))
#ggsave("figures/popSize_year.jpeg", units="in", width=11, height=7, dpi=300)
```

PPGE-USP population size by gender and position and year

```{r, echo=F, fig.height=10}
pop2 %>% mutate(ytext = M + F/2) %>%
  pivot_longer(7:8, names_to = "gender", values_to = "N") %>%
  mutate(category = fct_relevel(category, "student", "postdoc", "professor")) %>%
  ggplot(aes(x=as.factor(year),y=N, fill=gender)) + geom_col() + 
  facet_wrap(~category, scales="free_y",ncol=1)+
  ylab("N") + xlab("") +
   scale_fill_manual("gender", values = c("#b2abd2", "#fdb863")) +
  theme(text = element_text(size=18),
    axis.text.x = element_text(size=16, angle=45, hjust=1)) +
  geom_text(aes(x=as.factor(year), y=ytext,
                label=paste0(round(propFcat*100), "%") ))

#ggsave("figures/FIG_S1_popSize_positionYear.jpeg", units="in", width=7, height=10, dpi=300)
```


# Data description

Descriptive totals: From `r dim(data1)[1]` talks in 12 years, we used `r dim(data_all)[1]`  as we excluded special seminars and round tables. 

In total, `r paste0(round(((data_all %>% tabyl(gender))[2,3]*100)), "%")` of the talks were given by men (n=`r (data_all %>% tabyl(gender))[2,2]`) and `r paste0(round(((data_all %>% tabyl(gender))[1,3]*100)), "%")` by women (n=`r (data_all %>% tabyl(gender))[1,2]`).

```{r}
dim(data_all)
data_all %>% tabyl(gender) %>% kable()
```

# Speakers data

Proportion and number of male and female speakers per academic position.

```{r}
data_p <- data_all %>% filter(position_cat != "others")
data_p$position_cat <- fct_relevel(data_p$position_cat, "student", 
                                 "postdoc","professor")
```

Excluding talks from non-academic professionals.  N = `r dim(data_p)[1]`.

```{r, echo=F}
data_p %>% tabyl(position_cat, gender) %>% adorn_percentages("row") %>%
  adorn_pct_formatting(digits = 0) %>%
  adorn_ns() %>%
  kable(caption="Proportion and number (in parenthesis) of female (F) and male (M) speakers per academic position.")
```

```{r num speak, echo=F}
n.y <- data_p %>% tabyl(position_cat, gender) %>% as.data.frame()
fig1<- ggplot(data_p, aes(x=position_cat, fill=gender)) + geom_bar() + 
  ylab("Number of speakers") + xlab("") +
   scale_fill_manual("gender", values = c("#b2abd2", "#fdb863")) +
  scale_color_discrete(name="Gênero") +
  theme(text = element_text(size=18),
    axis.text.x = element_text(size=16)) +
 ggplot2::annotate("text", x=1:3, y=n.y$M+(n.y$`F`/2), size=5,
           label = c("52%", "43%","24%"))
fig1
#ggsave("figures/numberSpeakers_position.jpeg", units="in", width=7, height=4.5, dpi=300)
```

Variation in time.

```{r, echo=F}
prop.y <- data_p %>% tabyl(year,gender) %>%
  adorn_percentages("row") %>%
  adorn_pct_formatting(digits = 0) 
n.y <- data_p %>% tabyl(year,gender)

fig2 <- ggplot(data_p, aes(x=as.factor(year), fill=gender)) + geom_bar()  +
  ylab("Number of speakers") + xlab("") +
   scale_fill_manual("gender", values = c("#b2abd2", "#fdb863")) +
  theme(text=element_text(size=18),
    axis.text.x = element_text(size=16,angle = 45, hjust=1)) +
  geom_vline(xintercept = 10.5, linetype="dashed")+
  annotate("text", x=1:12, y=n.y$M+n.y$F/2,size=5,
           label = prop.y$F)
fig2
#ggsave("figures/numberSpeakers_year.jpeg", units="in", width=11, height=7, dpi=300)
```


```{r figure 1, echo=F, results='hide',fig.show='hide'}
fig1 + labs(tag = "a)") +
  fig2 + labs(tag = "b)") +
  plot_layout(widths = c(0.4,1), guides = "collect")
#ggsave("figures/FIG_1_numberSpeakers_position_year.jpeg", units="in", width=14, height=6, dpi=300)
```


### Origin of the speakers

```{r, echo=F}
origin <- paste(data_p$department, data_p$institute,data_p$university)
eco.ib <- sort(table(origin), decreasing = T)[1]
todo.ib <- length(grep("IB USP", origin, value=T))
```

There were `r sort(table(origin), decreasing = T)[1]` (`r round(eco.ib*100/dim(data_p)[1])`%) talks given by people from the PPGE population.

Including talks from the Institute of Biosciences, USP, there were `r todo.ib` (`r round(todo.ib*100/dim(data_p)[1])`%).



### Comparing proportions of female speakers in the seminar and in the PPGE population by position.

```{r,echo=F, fig.cap="Relationship between the proportion of females in the PPGE population in each category of academic position in each year (x axis) and the proportion of female speakers in the same category and year. Horizontal and vertical gray dashed lines are the 50% of each proportion and diagonal black dashed line indicates where the propotions are similar."}
propS <- data_p %>% group_by(year,position_cat,gender) %>% 
  count() %>% 
  pivot_wider(names_from = gender, values_from = n, values_fill = 0) %>%
  mutate(propFspeaker = F/(M+F)) 
proportions <- pop2 %>% dplyr::select(year, category, propFcat, F,M) %>% 
  rename("position_cat"="category") %>%
  left_join(propS, by=c("year", "position_cat")) %>%
  mutate(position_cat = fct_relevel(position_cat, "student", "postdoc", "professor"))
ggplot(proportions,aes(x=propFcat, y=propFspeaker, col=year)) +
  scale_color_gradient()+
  geom_point() +
  facet_wrap(~position_cat)+
  xlim(0,1) +
  xlab("Proportion of females in the population") +
  ylab("Proportion of females as speakers")+
  geom_abline(slope=1,intercept=0, linetype="dashed") +
  geom_hline(yintercept = 0.5, linetype="dashed", col="lightgray")+
  geom_vline(xintercept = 0.5, linetype="dashed", col="lightgray")
#ggsave("figures/propFemale_popXspeaker.jpeg", units="in", width=14, height=5, dpi=300)
```

# Audience data

Excluding talks from non-academic professionals.

Excluding seminars with more than one speaker:

```{r}
events <- data_p %>% count(id) %>% filter(n>1)
data_a <- data_p %>% filter(!id %in% events$id,
                        !is.na(audience_n))
```


There were `r length(data_a$audience_female!=NA)` talks with information about the number of attendees.

```{r}
data_a %>% group_by(position_cat,gender) %>% summarise(N=n(), 
                                                       min=min(audience_n),
                                                       mean=mean(audience_n),
                                                       sd = sd(audience_n),
                                                       median=median(audience_n),
                                                       max=max(audience_n)) %>%
  kable(digits=2)
```


```{r, echo=F, eval=F}
ggplot(data_a, aes(x=position_cat, y=audience_n, fill=gender)) +
  scale_fill_manual(values = c("#b2abd2", "#fdb863"))+
  xlab("Audience (N)")+
  geom_boxplot()
  #geom_violin(position = position_dodge(0.8)) +
  #geom_jitter(position=position_jitterdodge(0.2),shape=21)
```

```{r, echo=F}
library(ggbeeswarm)
# outra opção
ggplot(data_a, aes(x=position_cat, y=audience_n, fill=gender)) +
  scale_fill_manual(values = c("#b2abd2", "#fdb863"))+
  scale_color_manual(values = c("#b2abd2", "#fdb863"))+
  geom_violin(col="black") +
  geom_quasirandom(dodge.width = 0.9, shape=21)+
  stat_summary(fun.y=median, aes(ymin=..y.., ymax=..y..),geom='errorbar', 
               width=0.8, size=0.8, position = position_dodge(width = 0.9))+
  xlab("") + ylab("Audience (N)")
```


Looking for possible influence of the speakers origin (inside and outside PPGE) in the audience. 

```{r}
data_a$ppge <- ifelse(data_a$origin == "IB", "inside", "outside")
table(data_a$gender,data_a$ppge)
```

```{r, echo=F}
ggplot(data_a, aes(x=ppge, y=audience_n, fill=gender)) +
  scale_fill_manual(values = c("#b2abd2", "#fdb863"))+
  scale_color_manual(values = c("#b2abd2", "#fdb863"))+
  geom_violin(col="black") +
  geom_quasirandom(dodge.width = 0.9, shape=21)+
  stat_summary(fun.y=median, aes(ymin=..y.., ymax=..y..),geom='errorbar', 
               width=0.8, size=0.8, position = position_dodge(width = 0.9))+
  xlab("PPGE") + ylab("Audience (N)")
```

Looking for possible influence of the speaker origing (Brazil or abroad). 

```{r}
data_a$brazilian <- ifelse(data_a$country == "Brasil", "yes", "no")
table(data_a$gender,data_a$brazilian)
```

```{r, echo=F}
ggplot(data_a, aes(x=brazilian, y=audience_n, fill=gender)) +
  scale_fill_manual(values = c("#b2abd2", "#fdb863"))+
  scale_color_manual(values = c("#b2abd2", "#fdb863"))+
  geom_violin(col="black") +
  geom_quasirandom(dodge.width = 0.9, shape=21)+
  stat_summary(fun.y=median, aes(ymin=..y.., ymax=..y..),geom='errorbar', 
               width=0.8, size=0.8, position = position_dodge(width = 0.9))+
  xlab("Brazilian") + ylab("Audience (N)")
```



# Text analysis data

Excluding talks without title

```{r}
data_t <- data_all %>% filter(!is.na(title_english))
```

There were `r dim(data_t)[1]` talks with title information.


# Saving processed data

PPGE population data
```{r}
save(pop2, file="output/data_pop_ppge.Rdata")
```

Data for the proportion of female speakers
```{r}
save(data_p, file="output/data_proportion.Rdata")
```

Data for the audience of the talks
```{r}
save(data_a, file="output/data_audience.Rdata")
```

Data for the text analysis
```{r}
save(data_t, file="output/data_title.Rdata")
```
